refactor(settings): reduce Settings.tsx to a thin tab-shell over extracted tabs
