+++ app/renderer/src/routes/settings/TranscriptionTab.tsx
@@ -8,8 +8,10 @@ import {
   SelectValue,
 } from '@/components/ui/select';
 import {
+  useAutoSummarizeSetting,
   useKeepRecordingsSetting,
   useLanguageSetting,
+  useSetAutoSummarize,
   useSetKeepRecordings,
   useSetLanguage,
 } from '@/hooks/useSettings';
@@ -30,6 +32,8 @@ export function TranscriptionTab() {
   const setLanguage = useSetLanguage();
   const keepRecordings = useKeepRecordingsSetting();
   const setKeepRecordings = useSetKeepRecordings();
+  const autoSummarize = useAutoSummarizeSetting();
+  const setAutoSummarize = useSetAutoSummarize();
   const engineQuery = useTranscriptionEngine();
 
   const engine = engineQuery.data ?? 'parakeet';
@@ -79,6 +83,17 @@ export function TranscriptionTab() {
           checked={keepRecordings.data ?? false}
           onCheckedChange={(v) => setKeepRecordings.mutate(v)}
           disabled={keepRecordings.data === undefined}
+        />
+      </SettingRow>
+
+      <SettingRow
+        label="Generate notes automatically"
+        description="Summarise each recording right after transcription. Turn off to stop at a transcript and generate notes on demand."
+      >
+        <Switch
+          checked={autoSummarize.data ?? true}
+          onCheckedChange={(v) => setAutoSummarize.mutate(v)}
+          disabled={autoSummarize.data === undefined}
         />
       </SettingRow>
